Use uint32_t instead of size_t

+++ Source/TitaniumKit/src/Blob.cpp
@@ -59,7 +59,7 @@ namespace Titanium
 		// TODO image* methods!
 	}
 
-	size_t Blob::get_length() const TITANIUM_NOEXCEPT
+	uint32_t Blob::get_length() const TITANIUM_NOEXCEPT
 	{
 		return data_.size();
 	}
@@ -99,7 +99,7 @@ namespace Titanium
 		return path_;
 	}
 
-	size_t Blob::get_size() const TITANIUM_NOEXCEPT
+	uint32_t Blob::get_size() const TITANIUM_NOEXCEPT
 	{
 		if (type_ == Titanium::BlobModule::TYPE::IMAGE) {
 			return get_width() * get_height();
